+++ app/components/pages/index.js
@@ -141,6 +141,7 @@ Pages.propDescriptions = {
 	title: 'displayed title of the pages container',
 	bsSize: 'size of the pagination buttons (one of `default`, `lg`, `large`, `sm`, `small`, `xs`, or `xsmall`)',
 	height: 'the maximum height of the container. If an embedded page is taller, a vertical scrollbar is added',
+	voiceID: 'voice control identifier',
 	style: 'CSS inline styles',
 	onSelect: 'Function invoked when active change is changed. Receives the new active page index as a sole parameter'
 };